add abstract DungeonElement
