+++ code/core/src/basiselements/HUDElement.java
@@ -2,44 +2,32 @@ package basiselements;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import graphic.HUDPainter;
-import tools.Point;
 
-public abstract class HUDElement {
-    private SpriteBatch batch;
+public abstract class HUDElement extends DungeonElement {
     private HUDPainter painter;
+
     /**
      * A object that can be controlled by the <code>HUDController
      * </code>.
+     *
+     * @param painter Painter that draws this object
+     * @param batch   Batch to draw on
      */
     public HUDElement(HUDPainter painter, SpriteBatch batch) {
-        this.batch = batch;
+        super(batch);
         this.painter = painter;
     }
 
-    /** Will be executed every frame. */
-    public void update() {}
-
-    /** @return <code>true</code>, if this instance can be deleted; <code>false</code> otherwise */
-    public boolean removable() {
-        return false;
-    }
-
-    public SpriteBatch getBatch() {
-        return batch;
-    }
-
-    /** @return the exact position in the dungeon of this instance */
-    public abstract Point getPosition();
-
-    /** @return the (current) Texture-Path of the object */
-    public abstract String getTexturePath();
-
-    /** Each drawable should use this <code>Painter</code> to draw itself. */
+    /**
+     * Each drawable should use this <code>Painter</code> to draw itself.
+     */
     public HUDPainter getPainter() {
         return painter;
     }
 
-    /** Draws this instance on the batch. */
+    /**
+     * Draws this instance on the batch.
+     */
     public void draw() {
         getPainter().draw(getTexturePath(), getPosition(), getBatch());
     }
@@ -52,6 +40,7 @@ public abstract class HUDElement {
      */
     public void drawWithScaling(float xScaling, float yScaling) {
         getPainter()
-                .drawWithScaling(xScaling, yScaling, getTexturePath(), getPosition(), getBatch());
+            .drawWithScaling(xScaling, yScaling, getTexturePath(), getPosition(), getBatch());
     }
+
 }
