+++ code/core/src/graphic/HUDPainter.java
@@ -7,26 +7,30 @@ import textures.TextureMap;
 import tools.Constants;
 import tools.Point;
 
-/** Uses LibGDX to draw sprites on the various <code>SpriteBatch</code>es. */
+/**
+ * Uses LibGDX to draw sprites on the various <code>SpriteBatch</code>es.
+ */
 public class HUDPainter {
     private final TextureMap textureMap = new TextureMap();
 
-    /** Draws the instance based on its position. */
+    /**
+     * Draws the instance based on its position.
+     */
     public void draw(String texturePath, Point position, SpriteBatch batch) {
         Texture texture = textureMap.getTexture(texturePath);
         Sprite sprite = new Sprite(texture);
 
         // set up scaling of textures
         sprite.setSize(
-                texture.getWidth() / Constants.DEFAULT_ZOOM_FACTOR,
-                texture.getHeight() / Constants.DEFAULT_ZOOM_FACTOR);
+            texture.getWidth() / Constants.DEFAULT_ZOOM_FACTOR,
+            texture.getHeight() / Constants.DEFAULT_ZOOM_FACTOR);
 
         // where to draw the sprite
         sprite.setPosition(
-                position.x,
-                Constants.WINDOW_HEIGHT
-                        - position.y
-                        - texture.getHeight() / Constants.DEFAULT_ZOOM_FACTOR);
+            position.x,
+            Constants.WINDOW_HEIGHT
+                - position.y
+                - texture.getHeight() / Constants.DEFAULT_ZOOM_FACTOR);
 
         // need to be called before drawing
         batch.begin();
@@ -36,9 +40,11 @@ public class HUDPainter {
         batch.end();
     }
 
-    /** Draws the instance based on its position with default offset and specific scaling. */
+    /**
+     * Draws the instance based on its position with default offset and specific scaling.
+     */
     public void drawWithScaling(
-            float xScaling, float yScaling, String texturePath, Point position, SpriteBatch batch) {
+        float xScaling, float yScaling, String texturePath, Point position, SpriteBatch batch) {
         Texture texture = textureMap.getTexture(texturePath);
         Sprite sprite = new Sprite(texture);
 
@@ -47,7 +53,7 @@ public class HUDPainter {
 
         // where to draw the sprite
         sprite.setPosition(
-                position.x, Constants.WINDOW_HEIGHT - position.y - texture.getHeight() * yScaling);
+            position.x, Constants.WINDOW_HEIGHT - position.y - texture.getHeight() * yScaling);
 
         // need to be called before drawing
         batch.begin();
